rename seg_eval_pkg.py to CellSegmentationEvaluator.py and add CSE3D function
Changes to v1.5: Rename seg_eval_pkg.py to CellSegmentationEvaluator.py and reconcile differences with the 3D cell segmentation evaluation code in 3DCellComposer.  Add new function CSE3D as an alternative to "single_method_eval_3D" so that inputs can be nd-arrays rather than AICSImage structures.
