+++ README.md
@@ -1,7 +1,7 @@
 # Cell Segmentation Evaluator: evaluation of cell segmentation methods without reference segmentations
 Haoran Chen and Robert F. Murphy and Ted Zhang\
 Carnegie Mellon University\
-V1.4 Dec 11, 2023
+V1.5 January 31, 2024
 
 ## Cell segmentation evaluation approach
 This package implements an approach for cell segmentation evaluation (CSE) that does not rely upon comparison to annotations from humans. 
@@ -10,6 +10,27 @@ We calculated these metrics for 11 previously-described segmentation methods app
 Using principal component analysis to combine the metrics, we defined an overall cell segmentation quality score.
 
 We also enhanced this tool by defining similar and new metrics to support 3D cell segmentations.
+
+The package can be installed using
+```bash
+pip install CellSegmentationEvaluator
+```
+
+Then import the desired function, e.g., 
+```bash
+from CellSegmentationEvaluator import single_method_eval, single_method_eval3D, CSE3D
+```
+
+There are three main functions to evaluate an image and associated masks.  For providing inputs as AICSImage structures, use 
+```bash
+single_method_eval(imgpath, maskpath) #for 2D
+single_method_eval3D(imgpath, maskpath) #for 3D
+```
+Or for providing 3D images and masks as nd-arrays, use
+```bash
+CSE3D(img, mask, [PCAmodel], [threshimage], [voxel_size])
+```
+where PCAmodel defaults to "3Dv1.6", threshimage defaults to the same of all channels, and the voxel size is in cubic micrometers and defaults to 1.
 
 Reference:
 
@@ -92,6 +113,9 @@ This step runs the evaluation pipeline given the generated configuration file.
 For a more detailed introduction to segmentation quality metrics among other image quality metrics, please see
 [HuBMAP Image Quality Control Metrics](http://hubmap.scs.cmu.edu/wp-content/uploads/2021/09/HuBMAP-Image-Quality-Control-Metrics-v1.5.pdf)
 
+## Changes from v1.4 to v1.5
+
+Rename seg_eval_pkg.py to CellSegmentationEvaluator.py and reconcile differences with the 3D cell segmentation evaluation code in 3DCellComposer.  Add new function CSE3D as an alternative to "single_method_eval_3D" so that inputs can be nd-arrays rather than AICSImage structures.
 
 ## Contact
